Remove generated man-page
The nroff file, wmpinboard.1, is generated from the POD file,
wmpinboard.pod, so remove it from git.  Since it is listed in
`dist_man_MANS`, it will be built as a matter of course when the dist
tar-ball is created.  Update the Makefile.am to ensure it can be cleaned
up.